requirements.txt
streamlit==1.32.0
python-dotenv==1.0.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
openai>=1.0.0
faiss-cpu>=1.7.4
pypdf==4.1.0
requests==2.32.3
tiktoken>=0.5.0 